requirement.txt
shapefile==
python-geojson==
flask-socketio==4.1.0